backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
numpy==1.24.2
scipy==1.10.1
pandas==2.0.3
statsmodels==0.14.0 
matplotlib==3.7.2               
matplotlib-inline==0.1.6
seaborn==0.12.2
scikit-posthocs==0.8.0 
